requirements.txt
fastapi
uvicorn
httpx
python-dotenv
telethon
jinja2
